move stuff to util, tried so hard to get subcommands working...
